functional test problem relates to AutoMapper explained(?) but not fixed
AutoMapper loads typeInfo at start up for every type that is referenced
recursively down the assemblies.  Some .NRT framwork types are referenced
in the newtelligence assemblies but do not exist in any of the .net
core bcl and runtime assemblies.  An exception is thrown.

Big question: why does this happen when FunctionalTests is the startup
module but not when DasBlog.Web is the startup module?